+++ src/main/java/frc/robot/commands/AutonomousCommands/AutonomousTaxiCommand.java
@@ -2,13 +2,10 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutonomousCommands;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.AutnomousCommands.GoToCommand;
-import frc.robot.commands.AutnomousCommands.LookAtCommand;
-import frc.robot.commands.AutnomousCommands.SetPoseCommand;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.ShuffleboardSubsystem;
