+++ src/main/java/frc/robot/commands/AutonomousCommands/GoToCommand.java
@@ -2,41 +2,29 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.AutnomousCommands;
+package frc.robot.commands.AutonomousCommands;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.ShuffleboardSubsystem;
 
-public class LookAtCommand extends CommandBase {
+public class GoToCommand extends CommandBase {
+
   private final DriveSubsystem m_subsystem;
   private Pose2d m_target;
-  private boolean shuffleboard;
-  /** Creates a new LookAtCommand using the target position from shuffleboard. */
-  public LookAtCommand(DriveSubsystem subsystem) {
+  /** Creates a new GoToCommand. */
+  public GoToCommand(DriveSubsystem subsystem) {
     // Use addRequirements() here to declare subsystem dependencies.
     m_subsystem = subsystem;
     addRequirements(m_subsystem);
-    shuffleboard = true;
   }
-
-  /** Creates a new LookAtCommand using the supplied target position */
-  public LookAtCommand(DriveSubsystem subsystem, Pose2d target) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    m_subsystem = subsystem;
-    addRequirements(m_subsystem);
-    m_target = target;
-    shuffleboard = false;
-  }
-
   // Called when the command is initially scheduled.
+  
   @Override
   public void initialize() {
-    if (shuffleboard) {
-      m_target = ShuffleboardSubsystem.getTargetPose();
-    }
-    m_subsystem.lookAt(m_target);
+    m_target = ShuffleboardSubsystem.getTargetPose();
+    m_subsystem.setRobotPosition(m_target);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
