+++ src/main/java/frc/robot/commands/AutonomousCommands/SetPoseCommand.java
@@ -2,29 +2,29 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.AutnomousCommands;
+package frc.robot.commands.AutonomousCommands;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import edu.wpi.first.wpilibj2.command.Subsystem;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.ShuffleboardSubsystem;
 
-public class GoToCommand extends CommandBase {
-
+public class SetPoseCommand extends CommandBase {
   private final DriveSubsystem m_subsystem;
-  private Pose2d m_target;
-  /** Creates a new GoToCommand. */
-  public GoToCommand(DriveSubsystem subsystem) {
+  private Pose2d m_newPose;
+  /** Creates a new SetPoseCommand. */
+  public SetPoseCommand(DriveSubsystem subsystem) {
     // Use addRequirements() here to declare subsystem dependencies.
     m_subsystem = subsystem;
     addRequirements(m_subsystem);
   }
+
   // Called when the command is initially scheduled.
-  
   @Override
   public void initialize() {
-    m_target = ShuffleboardSubsystem.getTargetPose();
-    m_subsystem.setRobotPosition(m_target);
+    m_newPose = ShuffleboardSubsystem.getStartingPose();
+    m_subsystem.resetPose(m_newPose);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
@@ -38,6 +38,6 @@ public class GoToCommand extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return m_subsystem.atTargetPosition();
+    return true;
   }
 }
